refactor(Polls): split up poll details
